requirements.txt
Django==1.10.6
djangorestframework==3.6.2
requests==2.13.0